requirements.txt
pip~=24.1.2
attrs~=24.2.0
wheel~=0.43.0
pillow~=10.4.0
tornado~=6.4.1
Jinja2~=3.1.4
filelock~=3.15.4
setuptools~=72.1.0
packaging~=24.1
zipp~=3.20.0
typing_extensions~=4.12.2
build~=1.2.1
chromadb~=0.5.5
python-dotenv~=1.0.1
langchain~=0.2.12
mistralai~=1.0.1
streamlit~=1.37.1
beautifulsoup4
tiktoken
langchain-chroma
langchain-mistralai
langchain-community